website/requirements.txt
asgiref==3.4.1
Django==3.2.5
django-bootstrap-form==3.4
django-easy-select2==1.5.7
django-filter==2.4.0
django-widget-tweaks==1.4.8
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
